--- packages/hekireki/src/generator/mermaid-er/validator/remove-duplicate-relations.ts
@@ -1,9 +1,0 @@
-/**
- * Remove duplicate relation lines from an array of Mermaid ER diagram relations.
- *
- * @param relations - An array of relation lines (e.g., generated from `relationLine`).
- * @returns A new array with duplicates removed, preserving insertion order.
- */
-export function removeDuplicateRelations(relations: readonly string[]): readonly string[] {
-  return [...new Set(relations)]
-}
