+++ packages/hekireki/src/generator/valibot/validator/index.test.ts
@@ -4,12 +4,26 @@ import { isValibotDocument, isValibot } from '.'
 // Test run
 // pnpm vitest run ./src/generator/valibot/validator/index.test.ts
 
-describe('valibot validator exports', () => {
-  it('should export isValibotDocument', () => {
-    expect(typeof isValibotDocument).toBe('function')
+describe('validator', () => {
+  // isValibot
+  describe('isValibotDocument', () => {
+    it.concurrent('isValibotDocument Test', () => {
+      const result = isValibotDocument(`Unique identifier for the user
+@z.uuid()
+@v.pipe(v.string(), v.uuid())`)
+      const expected = ['Unique identifier for the user', '@z.uuid()']
+      expect(result).toStrictEqual(expected)
+    })
   })
 
-  it('should export isValibot', () => {
-    expect(typeof isValibot).toBe('function')
+  // isValibot
+  describe('isValibot', () => {
+    it.concurrent('isValibot Test', () => {
+      const result = isValibot(`Unique identifier for the user
+@z.uuid()
+@v.pipe(v.string(), v.uuid())`)
+      const expected = 'pipe(v.string(), v.uuid())'
+      expect(result).toStrictEqual(expected)
+    })
   })
 })
