--- packages/hekireki/src/generator/valibot/validator/is-valibot-document.ts
@@ -1,11 +1,0 @@
-const VALIDATION_PATTERNS = ['@v.'] as const
-
-export function isValibotDocument(documentation?: string): string[] {
-  if (!documentation) return []
-
-  return documentation
-    .split('\n')
-    .filter((line) => !VALIDATION_PATTERNS.some((pattern) => line.includes(pattern)))
-    .map((line) => line.trim())
-    .filter(Boolean)
-}
