--- packages/hekireki/src/generator/valibot/validator/is-valibot.ts
@@ -1,10 +1,0 @@
-/**
- * Check if the documentation is a Valibot validation
- * @param documentation
- * @returns string | null
- */
-export function isValibot(documentation?: string): string | null {
-  if (!documentation) return null
-  const match = documentation.match(/@v\.(.+?)(?:\n|$)/)
-  return match ? match[1].trim() : null
-}
